Added extra attributes:   Liveout for marking live assignments (deadcode.ml should respect it now)   StrAttr of string for generic attributes users may want to parse and print   ExnAttr of exn for generic attributes which can contain non-textual info
It would be nice if there were a single system that would have the advantages
of both StrAttr and ExnAttr, but I don't think there is a way to do that. We
could try to print ExnAttrs, but then we wouldn't be able to parse them back in.

While testing Liveout I am encountering other bugs, so I'm not entirely sure
it works right.



git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/trunk@3811 f55ac123-dd1f-49ad-9780-241693a2721e