some fixes to allow parsing of trace il.
I wanted to parse contexts using the grammar, but attributes were currently all written out as strings.  I couldn't find a way of applying the grammar to a string inside of the grammar, so I made the context attribute not a string.  Feel free to improve on this if there is a more elegant solution... I just couldn't think of one.


git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/branches/traces@4354 f55ac123-dd1f-49ad-9780-241693a2721e