Add some initial recursive descent stuff.
The code is currently in a utility for easier testing, but will probably
be moved under ocaml/ in an upcoming commit.  Right now it only follows
direct jumps, so it won't get to main from __libc_start_main yet.

It is also quite slow for things that lead to lot of instructions (like
__libc_start_main, where it takes almost 5 seconds to disassemble 33060
instructions).


git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/branches/recursive_descent@5256 f55ac123-dd1f-49ad-9780-241693a2721e