Got this !@#$@#$*&* bug in the CFG to Ast.prog converter.
The converter works by extending straightline traces.  For instance, if A .. B is a trace from A to B and C .. D is a trace from C to D, these can be merged to create a trace from A to D.  Unfortunately, the existing code for this does this when if A == C, which creates a cycle.  In the process of creating the cycle, we lose track of the "head" of the cycle, and we never print any of the blocks in the cycle.


git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/branches/traces+merging@5410 f55ac123-dd1f-49ad-9780-241693a2721e