Trace container format is "code complete".
However, there is a problem with google protobuffers.  I was hoping I would be able to unserialize an istream, and protobuffers would only consume as much data as was necessary, but this does not seem to be the case.  So, we will need to wrote the length of frames, put them in a buffer, and then call protobuffers on that buffer.  Annoying.


git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/branches/traces+piqi@5704 f55ac123-dd1f-49ad-9780-241693a2721e