Undo my previous tightening change, since I misunderstood the meaning of "pfx" in the Intel opcode map.
I did still tighten the 0xd6 opcode for movq to require the opsize prefix.

Many of the instructions here are probably wrong (e.g., we accept prefix/opcode combinations not in the manual), so at some point we should go through and try to fix this, but it's not as simple as looking through the opcode map, as I originally thought.


git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/trunk@5979 f55ac123-dd1f-49ad-9780-241693a2721e