Add reduced_havlak implementation
Unfortunately, this doesn't work - I think it exposed a flaw in the
reduced_havlak algorithm itself (I ran the algorithm by hand and encountered
the same infinite loop that broke the code). As such, the
"make_explicit_representation" (which Ramalingam claimed was 'straightforward')
is untested. I'm commiting this now although it *doesn't work* - be warned.

git-svn-id: https://kestrel.ece.cmu.edu/svn/bap/branches/djgraph@7331 f55ac123-dd1f-49ad-9780-241693a2721e