Changed handling of dead cjmps in aggressive deadcode eliminator.
Previously, we were resolving dead cjmps by instead jumping to the point
that can still reach BB_Exit. This resulted in failures where both branches
led to infinite loops. This was resolved by making BB_Exit live, thus
preventing the case where one branch leads to BB_Exit while the other
does not (as we would have a live cjmp in this case rather than a dead one).
Then, we can choose one of the cjmp branches arbitrarily, as no live
statements are dependent on it, so the branches will either eventually
converge or lead to only dead code. If both branches lead to dead infinite
loops, it does not matter which we choose.